ASoC: rockchip: i2s-tdm: Optimize TRCM on slave mode
TRCM require TX/RX enabled at the same time, or need the one
which provide clk enabled at first for master mode.

It is quite a different for slave mode which does not have
these restrictions, because the BCLK / LRCK are provided by
external master devices.

So, we just set the right clk path value on TRCM register on
stage probe and then drop the trcm value to make TX / RX work
independently.

Signed-off-by: Sugar Zhang <sugar.zhang@rock-chips.com>
Change-Id: I104f168d75787d79625ca7765a1bda1e77cdde12