phy: rockchip: inno-usb2: enable bvalid detect irq upon resume
The voltage domain of the usb2 phy grf interrupt mask
register usually belongs to the VD_LOGIC in most of
Rockchip SoCs. So if the VD_LOGIC is power off after
system suspend, the configuration of the usb2 phy grf
interrupt mask may lost when resume.

Test on RK3588 platform with micro usb2.0 interface,
if this case happened, the usb device failed to connect
to host, because the usb device depends on the bvalid
irq to start the enumeration.

This patch enable the bvalid detect irq upon resume
for otg port, and schedule the otg sm delayed work
if the bvalid is high which means that usb device has
connected to Host.

Change-Id: I29245ec4fc812e45eb3f52cc5c2c270b659a0cc6
Signed-off-by: William Wu <william.wu@rock-chips.com>